src/backend/requirements.txt
fastapi==0.115.4
starlette==0.40.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
aiofiles==23.2.1

# for the button forwarder script
requests==2.32.3
pyserial==3.5

